Added collision detection between entities
Each Entity struct now has a vector called "collisions". This vector is made up of Entity integer tuples, where the Entity is the other entity in the collision, and the integer is the side which is being collided (0 = top, 1 = right, 2 = bottom, 3 = left). This vector is updated whenever an entity's position is updated. Only the entity that moved has its collision updated.